Seperated based on responsibility
